Add files via upload
My friend and I did a deep dive into Oscar Health and designed a dossier, intended to be a one stop shop for all information to make an informed investment decision. All information is from publicly available content and opinions offered are our own. We are not licensed to give advice and this should not be seen as such. This is just our work.